+++ optimum/graphcore/__init__.py
@@ -51,6 +51,7 @@ from .models.roberta import (
 from .models.t5 import PipelinedT5ForConditionalGeneration
 from .models.vit import PipelinedViTForImageClassification
 from .models.wav2vec2 import PipelinedWav2Vec2ForPreTraining
+from .pipelines import IPUFillMaskPipeline, IPUTokenClassificationPipeline, pipeline
 from .trainer import IPUTrainer, IPUTrainerState
 from .trainer_seq2seq import IPUSeq2SeqTrainer
 from .training_args import IPUTrainingArguments
