+++ tests/pipelines/test_pipelines_image_classification.py
@@ -14,7 +14,7 @@
 
 import unittest
 
-from optimum.graphcore.pipelines import pipeline
+from optimum.graphcore import pipeline
 from transformers import MODEL_FOR_IMAGE_CLASSIFICATION_MAPPING, PreTrainedTokenizer, is_vision_available
 from transformers.pipelines import ImageClassificationPipeline
 from transformers.testing_utils import is_pipeline_test, nested_simplify, require_torch, require_vision, slow
